Kill react-maskedinput with 🔥
