+++ package.json
@@ -31,7 +31,6 @@
     "lodash": "^3.10.1",
     "moment": "^2.11.0",
     "react-day-picker": "^1.2.0",
-    "react-maskedinput": "iest/react-maskedinput",
     "react-motion": "^0.3.0",
     "react-slider": "^0.5.1",
     "react-tagsinput": "1.4.7"
